payment-microservice/payment-micro-service-app/pyproject.toml
[tool.poetry]
name = "payment-micro-service-app"
version = "0.1.0"
description = ""
authors = ["askat <askat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
uvicorn = "^0.31.1"
redis-om = "^0.3.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
